requirements.txt
pillow
numpy